+++ tests/tests/lib.rs
@@ -9,8 +9,8 @@ mod coins;
 mod contract;
 mod dap;
 mod debugger;
-mod deployment;
 mod fee_collection_contract;
+mod local_node;
 
 mod gas_price;
 mod health;
