+++ tests/tests/local_node.rs
@@ -7,11 +7,15 @@ use fuel_core::chain_config::{
     StateConfig,
     TableEncoding,
 };
-use fuel_core_types::fuel_tx::GasCosts;
+use fuel_core_types::{
+    blockchain::header::LATEST_STATE_TRANSITION_VERSION,
+    fuel_tx::GasCosts,
+};
 use fuel_core_upgradable_executor::WASM_BYTECODE;
+use test_helpers::fuel_core_driver::FuelCoreDriver;
 
 #[test]
-fn test_deployment_chainconfig() -> anyhow::Result<()> {
+fn local_chainconfig_validity() -> anyhow::Result<()> {
     let path = "./../bin/fuel-core/chainspec/local-testnet";
     let stored_snapshot = SnapshotMetadata::read(path).unwrap();
     let mut chain_config = ChainConfig::from_snapshot_metadata(&stored_snapshot)?;
@@ -26,8 +30,16 @@ fn test_deployment_chainconfig() -> anyhow::Result<()> {
 
     if env::var_os("OVERRIDE_CHAIN_CONFIGS").is_some() {
         chain_config.state_transition_bytecode = WASM_BYTECODE.to_vec();
+        chain_config.genesis_state_transition_version =
+            Some(LATEST_STATE_TRANSITION_VERSION);
         std::fs::remove_file(&stored_snapshot.chain_config)?;
         chain_config.write(&stored_snapshot.chain_config)?;
+    } else {
+        assert_eq!(
+            chain_config.genesis_state_transition_version,
+            Some(LATEST_STATE_TRANSITION_VERSION),
+            "Genesis state transition version should match the one in the deployment directory"
+        );
     }
 
     let temp_dir = tempfile::tempdir()?;
@@ -69,4 +81,27 @@ fn test_deployment_chainconfig() -> anyhow::Result<()> {
     );
 
     Ok(())
+}
+
+#[tokio::test]
+async fn start_local_node_without_any_arguments() {
+    // Given
+    let arguments = [];
+
+    // When
+    let service = FuelCoreDriver::spawn(&arguments).await.unwrap();
+
+    // Then
+    let block = service
+        .client
+        .block_by_height(0u32.into())
+        .await
+        .unwrap()
+        .unwrap();
+
+    assert_eq!(block.header.height, 0u32);
+    assert_eq!(
+        block.header.state_transition_bytecode_version,
+        LATEST_STATE_TRANSITION_VERSION
+    );
 }
